requirements_deploy.txt
# ── FangBlenny Bot — Railway requirements ────────────────────────
# Core bot deps (from original Termux setup)
requests>=2.31.0
websocket-client>=1.7.0
colorama>=0.4.6
python-dotenv>=1.0.0
blessed>=1.20.0

# API server
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0

# Supabase persistence
supabase>=2.4.0

# Optional — bot uses these if available
cryptography>=42.0.0
